Profile Update
Updated the users profile.
The user can now update their profile with their desired information.
The information they currently have is also displayed to the preview of their card and already initially displayed on the profile itself.

The structure of the retrieval is setup for the database side.
Resolve #77
Resolve #78
This solves 2 issues.